requirements.txt
marshmallow-sqlalchemy==0.26.1
oracledb==2.1.2
pandas==1.5.3